+++ src/lib/theme.tsx
@@ -1,0 +1,6 @@
+/**
+ * Windows XP design tokens and styled-components helpers.
+ */
+
+export { xpButtonStyles, xpScrollbarStyles, xpTitleBarStyles } from '../theme';
+export { COLORS, WINDOW_DEFAULTS, DESKTOP_DEFAULTS, TIME } from '../constants';